[fix]デバフ付与型の処理をBaseだけで完結するように
